Fixes issue #2236 - Consolidate session unit tests
