+++ core/impl/src/test/java/cloud/piranha/core/impl/tests/ServletRequestListenerTest.java
@@ -36,7 +36,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import cloud.piranha.core.api.WebApplication;
-import cloud.piranha.core.impl.DefaultHttpSessionManager;
 import cloud.piranha.core.impl.DefaultWebApplication;
 
 /**
@@ -59,7 +58,6 @@ class ServletRequestListenerTest {
     @BeforeEach
     void setUp() throws Exception {
         webApplication = new DefaultWebApplication();
-        webApplication.setHttpSessionManager(new DefaultHttpSessionManager());
     }
 
     /**
