+++ core/impl/src/test/java/cloud/piranha/core/impl/tests/ServletTest.java
@@ -45,7 +45,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import cloud.piranha.core.api.WebApplication;
-import cloud.piranha.core.impl.DefaultHttpSessionManager;
 import cloud.piranha.core.impl.DefaultWebApplication;
 
 /**
@@ -68,7 +67,6 @@ class ServletTest {
     @BeforeEach
     void setUp() throws Exception {
         webApplication = new DefaultWebApplication();
-        webApplication.setHttpSessionManager(new DefaultHttpSessionManager());
     }
 
     /**
