+++ core/servlet-api/src/test/java/jakarta/servlet/ServletRequestAttributeEventTest.java
@@ -27,8 +27,6 @@
  */
 package jakarta.servlet;
 
-import jakarta.servlet.ServletContext;
-import jakarta.servlet.ServletRequestAttributeEvent;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.Test;
