+++ core/servlet-api/src/test/java/jakarta/servlet/TestAsyncContext.java
@@ -27,13 +27,6 @@
  */
 package jakarta.servlet;
 
-import jakarta.servlet.ServletContext;
-import jakarta.servlet.AsyncListener;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.AsyncContext;
-import jakarta.servlet.ServletResponse;
-import jakarta.servlet.ServletRequest;
-
 /**
  * A Test AsyncContext.
  *
