+++ core/servlet-api/src/test/java/jakarta/servlet/TestServletRequest.java
@@ -27,16 +27,14 @@
  */
 package jakarta.servlet;
 
-import jakarta.servlet.ServletInputStream;
-import jakarta.servlet.AsyncContext;
-import jakarta.servlet.DispatcherType;
-import jakarta.servlet.ServletRequest;
-import jakarta.servlet.ServletRequestWrapper;
-
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Enumeration;
+import java.util.Locale;
+import java.util.Map;
 
 /**
  * A test Servlet request.
