+++ core/servlet-api/src/test/java/jakarta/servlet/http/HttpServletRequestWrapperTest.java
@@ -27,13 +27,11 @@
  */
 package jakarta.servlet.http;
 
+import java.util.Map;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.util.Map;
-
 import org.junit.jupiter.api.Test;
 
 /**
