+++ core/servlet-api/src/test/java/jakarta/servlet/http/HttpServletResponseWrapperTest.java
@@ -28,10 +28,14 @@
 package jakarta.servlet.http;
 
 import jakarta.servlet.TestServletOutputStream;
-
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * The JUnit tests for the HttpServletResponseWrapper class.
