+++ core/servlet-api/src/test/java/jakarta/servlet/http/HttpSessionBindingEventTest.java
@@ -27,10 +27,9 @@
  */
 package jakarta.servlet.http;
 
-import jakarta.servlet.http.HttpSessionBindingEvent;
-import jakarta.servlet.http.HttpSession;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * The JUnit tests for the HttpSessionBindingEventTest class.
