+++ core/servlet-api/src/test/java/jakarta/servlet/http/TestHttpServletResponse.java
@@ -27,14 +27,15 @@
  */
 package jakarta.servlet.http;
 
-import jakarta.servlet.http.HttpServletResponseWrapper;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletOutputStream;
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.util.*;
-
-import jakarta.servlet.ServletOutputStream;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
 
 /**
  * A Test HttpServletResponse.
